refactor: Fix cyclic imports in service modules
Extracted common types to service/types.rs:
- ActiveTab, ActiveViewer, AppState

Dependency flow is now unidirectional:
- types.rs ← explorer_state.rs, viewer_state.rs
- app_state.rs re-exports from types.rs for backward compatibility
