+++ src-tauri/src/service/app_state.rs
@@ -1,45 +1,11 @@
-//! アプリケーション状態管理の共通部分
+//! アプリケーション状態管理の共通ユーティリティ
 
-use notify::RecommendedWatcher;
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::sync::Arc;
 use tauri_plugin_dialog::DialogExt;
-use tokio::sync::{Mutex, RwLock};
 
 use crate::utils::file_utils::get_any_extensions;
 
-use super::explorer_state::ExplorerState;
-use super::viewer_state::ViewerState;
-
-// ========================================
-// 共通型定義
-// ========================================
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ActiveViewer {
-    pub label: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ActiveTab {
-    pub key: String,
-}
-
-// ========================================
-// アプリケーション状態
-// ========================================
-
-pub struct AppState {
-    pub count: Mutex<i32>,
-    pub active: Mutex<ActiveViewer>,
-    pub viewers: Mutex<Vec<ViewerState>>,
-    pub explorers: Mutex<Vec<ExplorerState>>,
-    /// ディレクトリ監視のwatcher管理 (path -> (watcher, 参照カウント))
-    pub watchers: Mutex<HashMap<String, (RecommendedWatcher, usize)>>,
-    /// サムネイルキャッシュ (folder_path -> thumbnail_path)
-    pub thumbnail_cache: Arc<RwLock<HashMap<String, String>>>,
-}
+// types.rs からの再エクスポート（後方互換性）
+pub use super::types::{ActiveTab, ActiveViewer, AppState};
 
 // ========================================
 // 共通ユーティリティ
